Add files via upload
Briefly summarize The Gaming Room client and their software requirements. Who was the client? What type of software did they want you to design?

The Gaming Room wants to develop a web-based game that serves multiple platforms based on their current game, Draw It or Lose It, which is currently available in an Android app only.

What about the process of working through a design document did you find helpful when developing the code?

It allows you to find issues and to work through the problems.  It is a very good tool to brainstorm and to start the project.

If you could choose one part of your work on these documents to revise, what would you pick? How would you improve it?

The evaluation section, I found  this difficult due to not having experience with the different operating platforms.  More research needs to be dome to understand the differences.

How did you interpret the user’s needs and implement them into your software design? Why is it so important to consider the user’s needs when designing?

Key needs to consider for the user usability, look, and feel. A user may continue to use the app if they don’t care for the look.

How did you approach designing software? What techniques or strategies would you use in the future to analyze and design a similar software application?

To design software, start vague and work towards being more complex. start small and work big.